requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.5.2
async-timeout==4.0.3
asyncpg==0.29.0
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
crypto==1.4.1
ecdsa==0.19.0
exceptiongroup==1.2.2
fastapi==0.115.2
frozenlist==1.4.1
h11==0.14.0
idna==3.10
multidict==6.1.0
Naked==0.1.32
passlib==1.7.4
propcache==0.2.0
pyasn1==0.6.1
pycrypto==2.6.1
pycryptodome==3.21.0
pydantic==2.9.2
pydantic_core==2.23.4
python-jose==3.3.0
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
shellescape==3.8.1
six==1.16.0
sniffio==1.3.1
starlette==0.39.2
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.31.1
yarl==1.15.2
